Update inventoryStore.js
added varun has hard coded fallback data. 892601 is used for the first time to login to the staff portal with details of varun (but here milind's is shown since its hardcoded in html). there was no previously borrowed items in the activity. only milind kumar staff and return all button.